pc-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
zeroconf==0.132.0
pystray==0.19.5
Pillow==10.4.0
pyperclip==1.9.0
plyer==2.1.0
python-multipart==0.0.9
aiofiles==24.1.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
